Added new 'entityAppointment' route and logic for creation
This commit adds a new endpoint at '/api/apts/entity/create' for the creation of appointment entities. An accompanying function, 'entityAppointment', was also added to the appointment.controller.js to handle these requests. Moreover, an additional function was implemented to appointment.service.js to generate usernames and support user creation for the appointment system.